Budget-Control App
Budget-Control App made with React and TypeScript as a part of Integrify frontend course. This app is made to help the client to organise their budget featuring such things as adding tha money to the account, planning the expences, saving money and so on. As a result of work on this tasks, the basic principles of working with React Components and React Hooks were studied and worked out.